# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CTRL API - Control Agent Economy Platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type TemplateCategory =
  (typeof TemplateCategory)[keyof typeof TemplateCategory];

export const TemplateCategory = {
  crypto: "crypto",
  ecommerce: "ecommerce",
  content: "content",
  saas: "saas",
} as const;

export interface Template {
  id: number;
  name: string;
  slug: string;
  description: string;
  category: TemplateCategory;
  agentCount: number;
  roomCount: number;
  rating: number;
  usageCount: number;
  isPublished: boolean;
  createdAt: string;
}

export type CreateTemplateBodyCategory =
  (typeof CreateTemplateBodyCategory)[keyof typeof CreateTemplateBodyCategory];

export const CreateTemplateBodyCategory = {
  crypto: "crypto",
  ecommerce: "ecommerce",
  content: "content",
  saas: "saas",
} as const;

export interface CreateTemplateBody {
  name: string;
  description: string;
  category: CreateTemplateBodyCategory;
  isPublished?: boolean;
}

export type StationStatus = (typeof StationStatus)[keyof typeof StationStatus];

export const StationStatus = {
  idle: "idle",
  running: "running",
  paused: "paused",
  completed: "completed",
} as const;

export interface Station {
  id: number;
  name: string;
  templateId: number;
  templateName: string;
  status: StationStatus;
  progress: number;
  agentCount: number;
  activeAgents: number;
  roomCount: number;
  tasksCompleted: number;
  tasksTotal: number;
  createdAt: string;
}

export interface CreateStationBody {
  name: string;
  templateId: number;
}

export type UpdateStationBodyStatus =
  (typeof UpdateStationBodyStatus)[keyof typeof UpdateStationBodyStatus];

export const UpdateStationBodyStatus = {
  idle: "idle",
  running: "running",
  paused: "paused",
  completed: "completed",
} as const;

export interface UpdateStationBody {
  name?: string;
  status?: UpdateStationBodyStatus;
}

export type RoomType = (typeof RoomType)[keyof typeof RoomType];

export const RoomType = {
  research: "research",
  development: "development",
  design: "design",
  marketing: "marketing",
  operations: "operations",
  analytics: "analytics",
} as const;

export type RoomStatus = (typeof RoomStatus)[keyof typeof RoomStatus];

export const RoomStatus = {
  idle: "idle",
  active: "active",
  busy: "busy",
} as const;

export interface Room {
  id: number;
  stationId: number;
  name: string;
  type: RoomType;
  status: RoomStatus;
  agentCount: number;
  tasksCompleted: number;
}

export type AgentRole = (typeof AgentRole)[keyof typeof AgentRole];

export const AgentRole = {
  research: "research",
  strategy: "strategy",
  builder: "builder",
  content: "content",
  growth: "growth",
  analytics: "analytics",
} as const;

export type AgentStatus = (typeof AgentStatus)[keyof typeof AgentStatus];

export const AgentStatus = {
  idle: "idle",
  working: "working",
  paused: "paused",
  offline: "offline",
} as const;

export interface Agent {
  id: number;
  stationId: number;
  roomId: number;
  name: string;
  role: AgentRole;
  status: AgentStatus;
  level: number;
  experience: number;
  tasksCompleted: number;
  currentTask?: string | null;
}

export type CreateRoomBodyType =
  (typeof CreateRoomBodyType)[keyof typeof CreateRoomBodyType];

export const CreateRoomBodyType = {
  research: "research",
  development: "development",
  design: "design",
  marketing: "marketing",
  operations: "operations",
  analytics: "analytics",
} as const;

export interface CreateRoomBody {
  name: string;
  type: CreateRoomBodyType;
}

export type CreateAgentBodyRole =
  (typeof CreateAgentBodyRole)[keyof typeof CreateAgentBodyRole];

export const CreateAgentBodyRole = {
  research: "research",
  strategy: "strategy",
  builder: "builder",
  content: "content",
  growth: "growth",
  analytics: "analytics",
} as const;

export interface CreateAgentBody {
  name: string;
  role: CreateAgentBodyRole;
  roomId: number;
}

export type UpdateAgentBodyStatus =
  (typeof UpdateAgentBodyStatus)[keyof typeof UpdateAgentBodyStatus];

export const UpdateAgentBodyStatus = {
  idle: "idle",
  working: "working",
  paused: "paused",
  offline: "offline",
} as const;

export interface UpdateAgentBody {
  status?: UpdateAgentBodyStatus;
  roomId?: number;
}

export type TaskStatus = (typeof TaskStatus)[keyof typeof TaskStatus];

export const TaskStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
  failed: "failed",
} as const;

export type TaskPriority = (typeof TaskPriority)[keyof typeof TaskPriority];

export const TaskPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface Task {
  id: number;
  agentId: number;
  title: string;
  description: string;
  status: TaskStatus;
  progress: number;
  priority: TaskPriority;
  createdAt: string;
  completedAt?: string | null;
}

export type CreateTaskBodyPriority =
  (typeof CreateTaskBodyPriority)[keyof typeof CreateTaskBodyPriority];

export const CreateTaskBodyPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface CreateTaskBody {
  title: string;
  description: string;
  priority: CreateTaskBodyPriority;
}

export type UpdateTaskBodyStatus =
  (typeof UpdateTaskBodyStatus)[keyof typeof UpdateTaskBodyStatus];

export const UpdateTaskBodyStatus = {
  pending: "pending",
  in_progress: "in_progress",
  completed: "completed",
  failed: "failed",
} as const;

export interface UpdateTaskBody {
  status?: UpdateTaskBodyStatus;
  progress?: number;
}

export interface DashboardSummary {
  totalStations: number;
  activeStations: number;
  totalAgents: number;
  activeAgents: number;
  tasksCompletedToday: number;
  totalTemplates: number;
  overallProgress: number;
}

export interface ActivityItem {
  id: number;
  agentName: string;
  agentRole: string;
  stationName: string;
  action: string;
  details: string;
  timestamp: string;
}

export interface AgentPerformance {
  role: string;
  tasksCompleted: number;
  avgProgress: number;
  agentCount: number;
}

export type ListTemplatesParams = {
  category?: ListTemplatesCategory;
};

export type ListTemplatesCategory =
  (typeof ListTemplatesCategory)[keyof typeof ListTemplatesCategory];

export const ListTemplatesCategory = {
  crypto: "crypto",
  ecommerce: "ecommerce",
  content: "content",
  saas: "saas",
  all: "all",
} as const;

export type GetRecentActivityParams = {
  limit?: number;
};